requirements.txt
asposestorage
Flask
Flask_Login
Flask_Migrate
flask_sqlalchemy
Pillow
Werkzeug
gunicorn
psycopg2
